refactor: 랜덤 로또 번호 생성 담당 클래스 추가
