+++ src/main/java/lotto/domain/LottoNumbers.java
@@ -1,16 +1,11 @@
 package lotto.domain;
 
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 public class LottoNumbers {
-    private static final int LOTTO_NUMBERS_SIZE = 6;
     private static final String LOTTO_NUMBERS_STRING_DELIMITER = ", ";
 
     private final List<LottoNumber> lottoNumbers;
@@ -24,28 +19,8 @@ public class LottoNumbers {
     }
 
     public static LottoNumbers ofRandom() {
-        List<LottoNumber> randomLottoNumbers = getRandomLottoNumbers();
-
+        List<LottoNumber> randomLottoNumbers = RandomLottoNumbersGenerator.generate();
         return new LottoNumbers(randomLottoNumbers);
-    }
-
-    private static List<LottoNumber> getRandomLottoNumbers() {
-        Set<LottoNumber> randomLottoNumbers = new HashSet<>();
-        while (randomLottoNumbers.size() < LOTTO_NUMBERS_SIZE) {
-            LottoNumber randomLottoNumber = LottoNumber.ofRandom();
-            randomLottoNumbers.add(randomLottoNumber);
-        }
-
-        return sort(toList(randomLottoNumbers));
-    }
-
-    private static List<LottoNumber> toList(Set<LottoNumber> lottoNumbers) {
-        return new ArrayList<>(lottoNumbers);
-    }
-
-    private static List<LottoNumber> sort(List<LottoNumber> lottoNumbers) {
-        lottoNumbers.sort(Comparator.comparingInt(LottoNumber::getLottoNumber));
-        return lottoNumbers;
     }
 
     private static List<LottoNumber> toLottoNumbers(String lottoNumbersString) {
